Crate minidump_processor

A parser for the minidump file format.

The minidump_processor module provides a parser for the minidump file format as produced by Microsoft's MinidumpWriteDump API and the Google Breakpad library.

The primary API for this module is the Minidump struct, which can be instantiated by calling the Minidump::read or Minidump::read_path methods.

Modules

minidump_format
Minidump structure definitions (autogenerated from C headers).

Structs

CallStack
A stack of StackFrames produced as a result of unwinding a thread.

Minidump
An index into the contents of a minidump.

MinidumpBreakpadInfo
Additional information about process state.

MinidumpContext
CPU context such as register states.

MinidumpException
Information about the exception that caused the minidump to be generated.

MinidumpMemory
A region of memory from the process that wrote the minidump.

MinidumpMiscInfo
Miscellaneous information about the process that wrote the minidump.

MinidumpModule
An executable or shared library loaded in the process at the time the Minidump was written.

MinidumpModuleList
A list of MinidumpModules contained in a Minidump.

MinidumpSystemInfo
Information about the system that generated the minidump.

MinidumpThread
The state of a thread from the process when the minidump was written.

MinidumpThreadList
A list of MinidumpThreads contained in a Minidump.

ProcessState
The state of a process as recorded by a Minidump.

StackFrame
A single stack frame produced from unwinding a thread's stack.

SystemInfo
Information about the system that produced a Minidump.

Enums

CPU
Known CPU types.

CallStackInfo
Information about the results of unwinding a thread's stack.

Error
Errors encountered while reading a Minidump.

MinidumpRawContext
The CPU-specific context structure.

OS
Known operating systems.

ProcessError
An error encountered during minidump processing.

Traits

Module
An executable or shared library loaded in a process.

Functions

process_minidump
Unwind all threads in dump and return a ProcessState.