Struct minidump_processor::ProcessState

pub struct ProcessState {
    pub time: DateTime<UTC>,
    pub process_create_time: DateTime<UTC>,
    pub crashed: bool,
    pub crash_reason: Option<String>,
    pub crash_address: Option<u64>,
    pub assertion: Option<String>,
}

The state of a process as recorded by a Minidump.

Fields

time

When the minidump was written.

process_create_time

When the process started.

crashed

true if the minidump was written in response to a process crash.

crash_reason

If the process crashed, a String describing the crash reason.

This is OS- and possibly CPU-specific. For example, "EXCEPTION_ACCESS_VIOLATION" (Windows), "EXC_BAD_ACCESS / KERN_INVALID_ADDRESS" (Mac OS X), "SIGSEGV" (other Unix).

crash_address

The memory address implicated in the crash.

If the process crashed, and if the crash reason implicates memory, this is the memory address that caused the crash. For data access errors this will be the data address that caused the fault. For code errors, this will be the address of the instruction that caused the fault.

assertion

A string describing an assertion that was hit, if present.

Methods

impl ProcessState

fn print<T: Write>(&self, _f: &mut T) -> Result<()>

Write a human-readable description of the process state to _f.

This is very verbose, it implements the output format used by minidump_stackwalk.