Struct minidump_processor::StackFrame

pub struct StackFrame {
    pub instruction: u64,
    pub module: Option<MinidumpModule>,
    pub function_name: Option<String>,
    pub function_base: Option<u64>,
    pub source_file_name: Option<String>,
    pub source_line: Option<usize>,
    pub source_line_base: Option<u64>,
    pub trust: FrameTrust,
}

A single stack frame produced from unwinding a thread's stack.

Fields

instruction

module

function_name

The function name, may be omitted if debug symbols are not available.

function_base

The start address of the function, may be omitted if debug symbols are not available.

source_file_name

The source file name, may be omitted if debug symbols are not available.

source_line

The (1-based) source line number, may be omitted if debug symbols are not available.

source_line_base

The start address of the source line, may be omitted if debug symbols are not available.

trust

Amount of trust the stack walker has in the instruction pointer of this frame.

Methods

impl StackFrame

fn return_address(&self) -> u64

Return the actual return address, as saved on the stack or in a register. See the comments for StackFrame::instruction for details.